refactor(state): minor improvements in folder org
