+++ src/components/KeyResult/Table/Body/Component.tsx
@@ -12,7 +12,7 @@ import { useRecoilStateLoadable, useRecoilValueLoadable } from 'recoil'
 import { KeyResult } from 'components/KeyResult'
 import logger from 'lib/logger'
 import { hasFetchedAllValues } from 'specifications'
-import { buildCustomSorter } from 'state/actions'
+import { buildCustomSorter } from 'state/actions/users'
 import { allKeyResults as allKeyResultsAtom } from 'state/recoil/keyResults/all'
 import { userKeyResultsCustomSorting as userCustomSortingAtom } from 'state/recoil/users/current/customSorting/keyResults'
 import { userID as userIDAtom } from 'state/recoil/users/current/id'
