+++ src/state/actions/api/index.ts
@@ -1,0 +1,5 @@
+export { default as getFromAPI } from './getFromAPI'
+
+export { default as fetchFromAPI } from './fetchFromAPI'
+
+export { default as patchFromAPI } from './patchFromAPI'
