+++ src/state/recoil/keyResults/all.ts
@@ -3,7 +3,7 @@ import { atom, selector } from 'recoil'
 import { PREFIX } from './constants'
 
 import { KeyResultsHashmap } from 'components/KeyResult'
-import { getFromAPI } from 'state/actions'
+import { getFromAPI } from 'state/actions/api'
 
 export const KEY = `${PREFIX}::ALL`
