+++ src/state/recoil/users/current/customSorting/keyResults.ts
@@ -3,7 +3,7 @@ import { atom, selector } from 'recoil'
 import { PREFIX } from './constants'
 
 import { CustomSorting } from 'components/User'
-import { getFromAPI } from 'state/actions'
+import { getFromAPI } from 'state/actions/api'
 import { userID as userIDAtom } from 'state/recoil/users/current/id'
 
 export const KEY = `${PREFIX}::KEY_RESULTS`
